large chunk of changes ported from Miguel's personal repo
