+++ Lib/fontbakery/profiles/adobefonts.py
@@ -181,7 +181,6 @@ SET_EXPLICIT_CHECKS = {
     # "com.google.fonts/check/STAT_strings",
     # "com.google.fonts/check/transformed_components",
     # ---
-    "com.adobe.fonts/check/freetype_rasterizer",  # IS_OVERRIDDEN
     "com.google.fonts/check/family/win_ascent_and_descent",  # IS_OVERRIDDEN
     "com.google.fonts/check/fontbakery_version",  # IS_OVERRIDDEN
     "com.google.fonts/check/name/trailing_spaces",  # IS_OVERRIDDEN
@@ -189,6 +188,7 @@ SET_EXPLICIT_CHECKS = {
     "com.google.fonts/check/valid_glyphnames",  # IS_OVERRIDDEN
     "com.google.fonts/check/whitespace_glyphs",  # IS_OVERRIDDEN
     # ---
+    "com.adobe.fonts/check/freetype_rasterizer",
     "com.adobe.fonts/check/sfnt_version",
     "com.google.fonts/check/family/single_directory",
     "com.google.fonts/check/family/vertical_metrics",
@@ -222,7 +222,6 @@ ADOBEFONTS_PROFILE_CHECKS = [
 ] + [c for c in SET_IMPORTED_CHECKS if c in SET_EXPLICIT_CHECKS]
 
 OVERRIDDEN_CHECKS = [
-    "com.adobe.fonts/check/freetype_rasterizer",
     "com.adobe.fonts/check/stat_has_axis_value_tables",
     "com.adobe.fonts/check/varfont/valid_default_instance_nameids",
     "com.fontwerk/check/inconsistencies_between_fvar_stat",
@@ -576,14 +575,6 @@ profile.check_log_override(
         ("descender", WARN, KEEP_ORIGINAL_MESSAGE),
         ("lineGap", WARN, KEEP_ORIGINAL_MESSAGE),
     ),
-)
-
-
-profile.check_log_override(
-    # From universal.py
-    "com.adobe.fonts/check/freetype_rasterizer",
-    overrides=(("freetype-not-installed", FAIL, KEEP_ORIGINAL_MESSAGE),),
-    reason="For Adobe, this check is very important and should never be skipped.",
 )
