+++ Lib/fontbakery/profiles/fontval.py
@@ -9,6 +9,20 @@ from fontbakery.message import Message
 from fontbakery.fonts_profile import profile_factory
 from .shared_conditions import is_cff, is_variable_font
 
+
+try:
+    import lxml
+except ImportError:
+    import sys
+
+    sys.exit(
+        "\nTo run the fontval profile, one needs to install\n"
+        "fontbakery with the 'fontval' extra, like this:\n"
+        "\n"
+        "python -m pip install -U 'fontbakery[fontval]'\n\n"
+    )
+
+
 profile_imports = [".shared_conditions"]
 profile = profile_factory(
     default_section=Section("Checks inherited from Microsoft Font Validator")
