+++ Lib/fontbakery/profiles/shaping.py
@@ -19,10 +19,12 @@ import textwrap
 from difflib import ndiff
 from pathlib import Path
 from fontbakery.callable import check, condition
-from fontbakery.checkrunner import FAIL, PASS, SKIP
+from fontbakery.checkrunner import FAIL, PASS, SKIP, WARN
 from fontbakery.fonts_profile import profile_factory
 from fontbakery.message import Message
 from fontbakery.section import Section
+from fontTools.unicodedata import ot_tag_to_script
+from ufo2ft.constants import INDIC_SCRIPTS, USE_SCRIPTS
 from vharfbuzz import Vharfbuzz, FakeBuffer
 from os.path import basename, relpath
 from stringbrewer import StringBrewer
@@ -30,13 +32,15 @@ from collidoscope import Collidoscope
 
 shaping_basedir = Path("qa", "shaping_tests")
 
-profile_imports = ()
+profile_imports = ((".", ("shared_conditions",)),)
 profile = profile_factory(default_section=Section("Shaping Checks"))
 
 SHAPING_PROFILE_CHECKS = [
     "com.google.fonts/check/shaping/regression",
     "com.google.fonts/check/shaping/forbidden",
     "com.google.fonts/check/shaping/collides",
+    "com.google.fonts/check/dotted_circle",
+    "com.google.fonts/check/soft_dotted",
 ]
 
 STYLESHEET = """
@@ -239,8 +243,8 @@ def run_a_set_of_shaping_tests(
         any differences.
         
         Shaping test suites should be written by the font engineer and referenced
-        in the fontbakery configuration file. For more information about write
-        shaping test files and how to configure fontbakery to read the shaping
+        in the FontBakery configuration file. For more information about write
+        shaping test files and how to configure FontBakery to read the shaping
         test suites, see https://simoncozens.github.io/tdd-for-otl/
     """,
     proposal="https://github.com/googlefonts/fontbakery/pull/3223",
@@ -252,7 +256,7 @@ def com_google_fonts_check_shaping_regression(config, ttFont):
         ttFont,
         run_shaping_regression,
         lambda test, configuration: "expectation" in test,
-        gereate_shaping_regression_report,
+        generate_shaping_regression_report,
     )
 
 
@@ -273,7 +277,7 @@ def run_shaping_regression(
         failed_shaping_tests.append((test, expectation, output_buf, output_serialized))
 
 
-def gereate_shaping_regression_report(vharfbuzz, shaping_file, failed_shaping_tests):
+def generate_shaping_regression_report(vharfbuzz, shaping_file, failed_shaping_tests):
     report_items = []
     header = f"{shaping_file}: Expected and actual shaping not matching"
     report_items.append(create_report_item(vharfbuzz, header, type="header"))
@@ -312,8 +316,8 @@ def gereate_shaping_regression_report(vharfbuzz, shaping_file, failed_shaping_te
         not be produced. (For example, .notdef glyphs, visible viramas, etc.)
         
         Shaping test suites should be written by the font engineer and referenced in
-        the Font Bakery configuration file. For more information about write shaping
-        test files and how to configure fontbakery to read the shaping test suites,
+        the FontBakery configuration file. For more information about write shaping
+        test files and how to configure FontBakery to read the shaping test suites,
         see https://simoncozens.github.io/tdd-for-otl/
     """,
     proposal="https://github.com/googlefonts/fontbakery/pull/3223",
@@ -376,8 +380,8 @@ def forbidden_glyph_test_results(vharfbuzz, shaping_file, failed_shaping_tests):
         suite and reports instances where the glyphs collide in unexpected ways.
         
         Shaping test suites should be written by the font engineer and referenced
-        in the fontbakery configuration file. For more information about write
-        shaping test files and how to configure fontbakery to read the shaping
+        in the FontBakery configuration file. For more information about write
+        shaping test files and how to configure FontBakery to read the shaping
         test suites, see https://simoncozens.github.io/tdd-for-otl/
     """,
     proposal="https://github.com/googlefonts/fontbakery/pull/3223",
@@ -463,6 +467,246 @@ def collides_glyph_test_results(vharfbuzz, shaping_file, failed_shaping_tests):
         )
         report_items.append(report_item)
     yield FAIL, Message("shaping-collides", msg + ".\n" + "\n".join(report_items))
+
+
+def is_complex_shaper_font(ttFont):
+    for table in ["GSUB", "GPOS"]:
+        if table not in ttFont:
+            continue
+        if not ttFont[table].table.ScriptList:
+            continue
+        for rec in ttFont[table].table.ScriptList.ScriptRecord:
+            script = ot_tag_to_script(rec.ScriptTag)
+            if script in USE_SCRIPTS or script in INDIC_SCRIPTS:
+                return True
+            if script in ["Khmr", "Mymr", "Hang"]:
+                return True
+    return False
+
+
+@check(
+    id="com.google.fonts/check/dotted_circle",
+    conditions=["is_ttf"],
+    severity=3,
+    rationale="""
+        The dotted circle character (U+25CC) is inserted by shaping engines before
+        mark glyphs which do not have an associated base, especially in the context
+        of broken syllabic clusters.
+
+        For fonts containing combining marks, it is recommended that the dotted circle
+        character be included so that these isolated marks can be displayed properly;
+        for fonts supporting complex scripts, this should be considered mandatory.
+
+        Additionally, when a dotted circle glyph is present, it should be able to
+        display all marks correctly, meaning that it should contain anchors for all
+        attaching marks.
+    """,
+    proposal="https://github.com/googlefonts/fontbakery/issues/3600",
+)
+def com_google_fonts_check_dotted_circle(ttFont, config):
+    """Ensure dotted circle glyph is present and can attach marks."""
+    from fontbakery.utils import bullet_list, iterate_lookup_list_with_extensions
+
+    mark_glyphs = []
+    if (
+        "GDEF" in ttFont
+        and hasattr(ttFont["GDEF"].table, "GlyphClassDef")
+        and hasattr(ttFont["GDEF"].table.GlyphClassDef, "classDefs")
+    ):
+        mark_glyphs = [
+            k for k, v in ttFont["GDEF"].table.GlyphClassDef.classDefs.items() if v == 3
+        ]
+
+    # Only check for encoded
+    mark_glyphs = set(mark_glyphs) & set(ttFont.getBestCmap().values())
+    nonspacing_mark_glyphs = [g for g in mark_glyphs if ttFont["hmtx"][g][0] == 0]
+
+    if not nonspacing_mark_glyphs:
+        yield SKIP, "Font has no nonspacing mark glyphs."
+        return
+
+    if 0x25CC not in ttFont.getBestCmap():
+        # How bad is this?
+        if is_complex_shaper_font(ttFont):
+            yield FAIL, Message(
+                "missing-dotted-circle-complex",
+                "No dotted circle glyph present" "and font uses a complex shaper",
+            )
+        else:
+            yield WARN, Message(
+                "missing-dotted-circle", "No dotted circle glyph present"
+            )
+        return
+
+    # Check they all attach to dotted circle
+    # if they attach to something else
+    dotted_circle = ttFont.getBestCmap()[0x25CC]
+    attachments = {dotted_circle: []}
+    does_attach = {}
+
+    def find_mark_base(lookup, attachments):
+        if lookup.LookupType == 4:
+            # Assume all-to-all
+            for st in lookup.SubTable:
+                for base in st.BaseCoverage.glyphs:
+                    for mark in st.MarkCoverage.glyphs:
+                        attachments.setdefault(base, []).append(mark)
+                        does_attach[mark] = True
+
+    iterate_lookup_list_with_extensions(ttFont, "GPOS", find_mark_base, attachments)
+
+    unattached = []
+    for g in nonspacing_mark_glyphs:
+        if g in does_attach and g not in attachments[dotted_circle]:
+            unattached.append(g)
+
+    if unattached:
+        yield FAIL, Message(
+            "unattached-dotted-circle-marks",
+            f"The following glyphs could not be attached"
+            f" to the dotted circle glyph:\n\n"
+            f"{bullet_list(config, sorted(unattached))}",
+        )
+    else:
+        yield PASS, "All marks were anchored to dotted circle"
+
+
+@check(
+    id="com.google.fonts/check/soft_dotted",
+    severity=3,
+    rationale="""
+        An accent placed on characters with a "soft dot", like i or j, causes
+        the dot to disappear.
+        An explicit dot above can be added where required.
+        See "Diacritics on i and j" in Section 7.1, "Latin" in The Unicode Standard.
+
+        Characters with the Soft_Dotted property are listed in
+        https://www.unicode.org/Public/UCD/latest/ucd/PropList.txt
+
+        See also:
+        https://googlefonts.github.io/gf-guide/diacritics.html#soft-dotted-glyphs
+    """,
+    proposal="https://github.com/googlefonts/fontbakery/issues/4059",
+)
+def com_google_fonts_check_soft_dotted(ttFont):
+    """Ensure soft_dotted characters lose their dot when combined with marks that
+    replace the dot."""
+    import itertools
+    from beziers.path import BezierPath
+    from fontTools import unicodedata
+
+    cmap = ttFont["cmap"].getBestCmap()
+
+    # Soft dotted strings know to be used in orthographies.
+    ortho_soft_dotted_strings = set(
+        "i̋ i̍ i᷆ i᷇ i̓ i̊ i̐ ɨ́ ɨ̀ ɨ̂ ɨ̋ ɨ̏ ɨ̌ ɨ̄ ɨ̃ ɨ̈ ɨ̧́ ɨ̧̀ ɨ̧̂ ɨ̧̌ ɨ̱́ ɨ̱̀ ɨ̱̈ į́ į̀ į̂ į̄ į̄́ į̄̀ į̄̂ į̄̌ į̃ į̌ ị́ ị̀ ị̂ "
+        "ị̄ ị̃ ḭ́ ḭ̀ ḭ̄ j́ j̀ j̄ j̑ j̃ j̈ і́".split()
+    )
+    # Characters with Soft_Dotted property in Unicode.
+    soft_dotted_chars = set(
+        ord(c) for c in "iⅈ𝐢𝑖𝒊𝒾𝓲𝔦𝕚𝖎𝗂𝗶𝘪𝙞𝚒ⁱᵢįịḭɨᶤ𝼚ᶖjⅉ𝐣𝑗𝒋𝒿𝓳𝔧𝕛𝖏𝗃𝗷𝘫𝙟𝚓ʲⱼɉʝᶨϳіј"
+    ) & set(cmap.keys())
+    # Only check above marks used with Latin, Greek, Cyrillic scripts.
+    mark_above_chars = set(
+        (
+            c
+            for c in cmap.keys()
+            if unicodedata.combining(chr(c)) == 230
+            and unicodedata.block(chr(c)).startswith(
+                ("Combining Diacritical Marks", "Cyrillic")
+            )
+        )
+    )
+    # Only check non above marks used with Latin, Grek, Cyrillic scripts
+    # that are reordered before the above marks
+    mark_non_above_chars = set(
+        c
+        for c in cmap.keys()
+        if unicodedata.combining(chr(c)) < 230
+        and unicodedata.block(chr(c)).startswith("Combining Diacritical Marks")
+    )
+    # Skip when no characters to test with
+    if not soft_dotted_chars or not mark_above_chars:
+        yield SKIP, "Font has no soft dotted characters or no mark above characters."
+        return
+
+    # Collect outlines to skip fonts where i and dotlessi are the same,
+    # or i and I are the same.
+    outlines_dict = {
+        codepoint: BezierPath.fromFonttoolsGlyph(ttFont, glyphname)
+        for codepoint, glyphname in cmap.items()
+        if codepoint in [ord("i"), ord("I"), ord("ı")]
+    }
+    unclear = False
+    if ord("i") in cmap.keys() and ord("I") in cmap.keys():
+        if len(outlines_dict[ord("i")]) == len(outlines_dict[ord("I")]):
+            unclear = True
+    if not unclear and ord("i") in cmap.keys() and ord("ı") in cmap.keys():
+        if len(outlines_dict[ord("i")]) == len(outlines_dict[ord("ı")]):
+            unclear = True
+    if unclear:
+        yield SKIP, (
+            "It is not clear if the soft dotted" " characters have glyphs with dots."
+        )
+        return
+
+    # Use harfbuzz to check if soft dotted glyphs are substituted
+    filename = ttFont.reader.file.name
+    vharfbuzz = Vharfbuzz(filename)
+    fail_unchanged_strings = []
+    warn_unchanged_strings = []
+    for sequence in sorted(
+        itertools.product(
+            soft_dotted_chars,
+            # add "" to add cases without non above marks
+            mark_non_above_chars.union(set((0,))),
+            mark_above_chars,
+        )
+    ):
+        soft, non_above, above = sequence
+        if non_above:
+            unchanged = f"{cmap[soft]}|{cmap[non_above]}|{cmap[above]}"
+            text = chr(soft) + chr(non_above) + chr(above)
+        else:
+            unchanged = f"{cmap[soft]}|{cmap[above]}"
+            text = chr(soft) + chr(above)
+
+        # Only check a few strings that we WARN about.
+        if text not in ortho_soft_dotted_strings and len(warn_unchanged_strings) >= 20:
+            continue
+
+        buf = vharfbuzz.shape(text)
+        output = vharfbuzz.serialize_buf(buf, glyphsonly=True)
+        if output == unchanged:
+            if text in ortho_soft_dotted_strings:
+                fail_unchanged_strings.append(text)
+            else:
+                warn_unchanged_strings.append(text)
+
+    message = ""
+    if fail_unchanged_strings:
+        message += (
+            f"The dot of soft dotted characters used in orthographies "
+            f"must disappear in the following strings: "
+            f"{' '.join(fail_unchanged_strings)}"
+        )
+    if warn_unchanged_strings:
+        if message:
+            message += "\n\n"
+        message += (
+            f"The dot of soft dotted characters should disappear in "
+            f"other cases, for example: "
+            f"{' '.join(warn_unchanged_strings)}"
+        )
+    if fail_unchanged_strings:
+        yield FAIL, Message("soft-dotted", message)
+    elif warn_unchanged_strings:
+        yield WARN, Message("soft-dotted", message)
+    else:
+        yield PASS, (
+            "All soft dotted characters seem to lose their dot when "
+            "combined with a mark above."
+        )
 
 
 profile.auto_register(globals())
