+++ pyproject.toml
@@ -30,6 +30,7 @@ disable = [
 	"abstract-method",
 	"arguments-renamed",
 	"attribute-defined-outside-init",
+	"bare-except",
 	"broad-except",
 	"broad-exception-raised",
 	"c-extension-no-member",
@@ -55,20 +56,26 @@ disable = [
 	"invalid-name",
 	"keyword-arg-before-vararg",
 	"len-as-condition",
+	"line-too-long",
 	"logging-format-interpolation",
 	"logging-fstring-interpolation",
 	"logging-not-lazy",
 	"missing-docstring",
 	"missing-timeout",
+	"modified-iterating-dict",
 	"multiple-statements",
 	"no-else-raise",
 	"no-else-return",
 	"no-member",
 	"no-name-in-module",
+	"not-callable",
 	"pointless-statement",
+	"pointless-string-statement",
 	"protected-access",
 	"raise-missing-from",
+	"redefined-builtin",
 	"redefined-outer-name",
+	"reimported",
 	"singleton-comparison",
 	"subprocess-run-check",
 	"superfluous-parens",
@@ -84,15 +91,21 @@ disable = [
 	"too-many-public-methods",
 	"too-many-return-statements",
 	"too-many-statements",
+	"trailing-newlines",
+	"trailing-whitespace",
 	"undefined-loop-variable",
+	"undefined-variable",
 	"unexpected-keyword-arg",
+	"ungrouped-imports",
 	"unidiomatic-typecheck",
 	"unnecessary-comprehension",
 	"unnecessary-dict-index-lookup",
 	"unnecessary-pass",
 	"unspecified-encoding",
 	"unsubscriptable-object",
+	"unsupported-membership-test",
 	"unused-argument",
+	"unused-import",
 	"unused-variable",
 	"use-a-generator",
 	"use-dict-literal",
@@ -100,5 +113,9 @@ disable = [
 	"use-list-literal",
 	"use-sequence-for-iteration",
 	"used-before-assignment",
+	"useless-object-inheritance",
+	"useless-return",
 	"using-constant-test",
+	"wrong-import-order",
+	"wrong-import-position",
 ]
