+++ tests/profiles/shaping_test.py
@@ -3,14 +3,15 @@ import json
 import os
 import tempfile
 
-from fontbakery.checkrunner import FAIL
+from fontbakery.checkrunner import FAIL, WARN
 from fontbakery.codetesting import (
     assert_PASS,
+    assert_SKIP,
     assert_results_contain,
     CheckTester,
     TEST_FILE,
 )
-from fontbakery.profiles import universal as universal_profile
+from fontbakery.profiles import shaping as shaping_profile
 
 
 def wrap_args(config, font):
@@ -26,7 +27,7 @@ def wrap_args(config, font):
 
 def test_check_shaping_regression():
     """Check that we can test shaping against expectations."""
-    check = CheckTester(universal_profile, "com.google.fonts/check/shaping/regression")
+    check = CheckTester(shaping_profile, "com.google.fonts/check/shaping/regression")
 
     shaping_test = {
         "configuration": {},
@@ -52,7 +53,7 @@ def test_check_shaping_regression():
 
 def test_check_shaping_forbidden():
     """Check that we can test for forbidden glyphs in output."""
-    check = CheckTester(universal_profile, "com.google.fonts/check/shaping/forbidden")
+    check = CheckTester(shaping_profile, "com.google.fonts/check/shaping/forbidden")
 
     shaping_test = {
         "configuration": {"forbidden_glyphs": [".notdef"]},
@@ -78,7 +79,7 @@ def test_check_shaping_forbidden():
 
 def test_check_shaping_collides():
     """Check that we can test for colliding glyphs in output."""
-    check = CheckTester(universal_profile, "com.google.fonts/check/shaping/collides")
+    check = CheckTester(shaping_profile, "com.google.fonts/check/shaping/collides")
 
     shaping_test = {
         "configuration": {"collidoscope": {"area": 0, "bases": True, "marks": True}},
@@ -100,3 +101,39 @@ def test_check_shaping_collides():
             "shaping-collides",
             "ïï collides in Nunito",
         )
+
+
+def test_check_dotted_circle():
+    """Ensure dotted circle glyph is present and can attach marks."""
+    check = CheckTester(shaping_profile, "com.google.fonts/check/dotted_circle")
+
+    font = TEST_FILE("mada/Mada-Regular.ttf")
+    assert_PASS(check(font), "with a good font...")
+
+    font = TEST_FILE("cabin/Cabin-Regular.ttf")
+    assert_results_contain(check(font), WARN, "missing-dotted-circle")
+
+    font = TEST_FILE("broken_markazitext/MarkaziText-VF.ttf")
+    assert_results_contain(check(font), FAIL, "unattached-dotted-circle-marks")
+
+
+def test_check_soft_dotted():
+    """Check if font substitues soft dotted glyphs
+    when combined with top marks."""
+    check = CheckTester(shaping_profile, "com.google.fonts/check/soft_dotted")
+
+    font = TEST_FILE("abeezee/ABeeZee-Regular.ttf")
+    msg = assert_results_contain(check(font), WARN, "soft-dotted")
+    assert "The dot of soft dotted characters used in orthographies" not in msg
+    assert "The dot of soft dotted characters should disappear" in msg
+
+    font = TEST_FILE("cabin/Cabin-Regular.ttf")
+    msg = assert_results_contain(check(font), FAIL, "soft-dotted")
+    assert "The dot of soft dotted characters used in orthographies" in msg
+    assert "The dot of soft dotted characters should disappear" in msg
+
+    font = TEST_FILE("akshar/Akshar[wght].ttf")
+    assert_PASS(check(font), "All soft dotted characters seem to lose their dot ...")
+
+    font = TEST_FILE("rosarivo/Rosarivo-Regular.ttf")
+    assert_SKIP(check(font), "It is not clear if soft dotted characters ...")
